HAWKULAR-127 rename table filter directive to Directive. Move server status to its own file. Use ES6 fat arrow instead function(). Fix URL and App list redirection.
